requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
boto3==1.34.113
botocore==1.34.113
Brotli==1.1.0
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
cryptography==42.0.7
cssselect==1.3.0
cycler==0.12.1
dash==3.0.0
distro==1.9.0
et-xmlfile==1.1.0
ExifRead==3.0.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
firebase-admin==6.5.0
Flask==3.0.2
Flask-SocketIO==5.5.1
fonttools==4.56.0
google-ads==25.0.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
gspread==6.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.6
importlib_metadata==8.6.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.9.0
jmespath==1.0.1
joblib==1.3.2
kaitaistruct==0.10
kiwisolver==1.4.8
lxml==5.3.1
lxml_html_clean==0.4.2
MarkupSafe==2.1.5
matplotlib==3.10.1
msgpack==1.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.74.0
opencv-python==4.9.0.80
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.1
pillow==10.2.0
plotly==6.0.1
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyexiv2==2.5.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2024.1
PyYAML==6.0.2
regex==2023.12.25
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.20.0
selenium-wire==5.1.0
setuptools==75.2.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.2
StrEnum==0.4.15
stringcase==1.2.0
threadpoolctl==3.3.0
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
xlrd==2.0.1
zipp==3.21.0
zstandard==0.23.0
